feat: remove slide out mobile menu, add ecosystem link and tweak typography
