+++ apps/marginfi-v2-ui/src/components/mobile/MobileNavbar/MobileNavbar.tsx
@@ -1,48 +1,66 @@
-import { FC, useEffect, useMemo } from "react";
+import React from "react";
+
 import Link from "next/link";
-import { useMrgnlendStore, useUiStore, useUserProfileStore } from "~/store";
 import { useRouter } from "next/router";
+import { TablerIconsProps } from "@tabler/icons-react";
+
+import { useMrgnlendStore, useUiStore, useUserProfileStore } from "~/store";
 import { useFirebaseAccount } from "~/hooks/useFirebaseAccount";
 import { useWalletContext } from "~/hooks/useWalletContext";
 import { useOs } from "~/hooks/useOs";
 import { cn } from "~/utils/themeUtils";
-import { MenuModal } from "./MenuModal";
-import { ORDERED_MOBILE_NAVBAR_LINKS } from "~/config/navigationLinks";
-import { useSwipeGesture } from "~/hooks/useSwipeGesture";
 
-const MobileNavbar: FC = () => {
+import { IconBuildingBank, IconCoinOff, IconMoneybag, IconWorld } from "~/components/ui/icons";
+
+export interface NavLinkInfo {
+  href: string;
+  alt: string;
+  label: string;
+  Icon: (props: TablerIconsProps) => JSX.Element;
+}
+
+export const mobileLinks: NavLinkInfo[] = [
+  {
+    href: "/",
+    alt: "bank icon",
+    label: "lend",
+    Icon: IconBuildingBank,
+  },
+  {
+    href: "/stake",
+    alt: "mint icon",
+    label: "mint",
+    Icon: IconCoinOff,
+  },
+
+  {
+    href: "/portfolio",
+    alt: "portfolio icon",
+    label: "portfolio",
+    Icon: IconMoneybag,
+  },
+  {
+    href: "/ecosystem",
+    alt: "world icon",
+    label: "ecosystem",
+    Icon: IconWorld,
+  },
+];
+
+const MobileNavbar = () => {
   useFirebaseAccount();
 
-  const { walletAddress } = useWalletContext();
   const router = useRouter();
   const [isMenuModalOpen, setIsMenuModalOpen, isActionBoxInputFocussed] = useUiStore((state) => [
     state.isMenuDrawerOpen,
     state.setIsMenuDrawerOpen,
     state.isActionBoxInputFocussed,
   ]);
-  const [fetchPoints] = useUserProfileStore((state) => [state.fetchPoints]);
-  const [lendUserDataFetched, resetLendUserData] = useMrgnlendStore((state) => [
-    state.userDataFetched,
-    state.resetUserData,
-  ]);
 
   const { isIOS } = useOs();
 
-  useSwipeGesture(() => setIsMenuModalOpen(true));
-
-  useEffect(() => {
-    if (!walletAddress) return;
-    fetchPoints(walletAddress.toBase58()).catch(console.error);
-  }, [fetchPoints, walletAddress]);
-
-  useEffect(() => {
-    if (!walletAddress && lendUserDataFetched) {
-      resetLendUserData();
-    }
-  }, [walletAddress, lendUserDataFetched, resetLendUserData]);
-
-  const activeLink = useMemo(() => {
-    const activeLinkIndex = ORDERED_MOBILE_NAVBAR_LINKS.findIndex((link) => link.href === router.pathname);
+  const activeLink = React.useMemo(() => {
+    const activeLinkIndex = mobileLinks.findIndex((link) => link.href === router.pathname);
     return activeLinkIndex >= 0 ? `link${activeLinkIndex}` : "linknone";
   }, [router.pathname]);
 
@@ -51,8 +69,8 @@ const MobileNavbar: FC = () => {
   return (
     <footer>
       <nav className="fixed w-full bottom-0 z-50 bg-[#0F1111]">
-        <div className="h-full w-full text-sm font-[500] text-[#868E95] z-50 flex justify-around relative lg:gap-8">
-          {ORDERED_MOBILE_NAVBAR_LINKS.map((linkInfo, index) => {
+        <div className="h-full w-full text-xs font-normal text-[#868E95] z-50 flex justify-around relative lg:gap-8">
+          {mobileLinks.map((linkInfo, index) => {
             const isActive = activeLink === `link${index}`;
             return (
               <Link
@@ -60,24 +78,18 @@ const MobileNavbar: FC = () => {
                 onClick={() => linkInfo.label === "more" && setIsMenuModalOpen(true)}
                 href={linkInfo.href}
                 className={cn(
-                  "w-1/4 pt-3 flex flex-col items-center",
-                  isIOS ? "pb-7" : "pb-3",
-                  isActive ? "text-[#DCE85D]" : "text-[#999]"
+                  "w-1/4 py-2.5 flex flex-col gap-1 items-center border-t border-border",
+                  isIOS && "pb-4",
+                  isActive ? "text-chartreuse" : "text-[#999]"
                 )}
               >
                 <linkInfo.Icon />
-                <div className={`font-aeonik font-[400] ${isActive ? "text-[#DCE85D]" : "text-[#999]"}`}>
-                  {linkInfo.label}
-                </div>
+                <div className={`font-aeonik ${isActive ? "text-[#DCE85D]" : "text-[#999]"}`}>{linkInfo.label}</div>
               </Link>
             );
           })}
-
-          <div className={`w-full absolute top-[1px] border-t-[1px] border-[#333]`} />
-          <div className={`border-slider ${activeLink}`} />
         </div>
       </nav>
-      <MenuModal isOpen={isMenuModalOpen} handleClose={() => setIsMenuModalOpen(false)} />
     </footer>
   );
 };
